fix(shared): reuse favicon privacy rules across clients
